# docs/openapi.json
{
  "components": {
    "schemas": {
      "Answer": {
        "properties": {
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "claim_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Claim Ids",
            "type": "array"
          },
          "community_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Community Ids",
            "type": "array"
          },
          "entity_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Entity Ids",
            "type": "array"
          }
        },
        "required": [
          "answer",
          "citations",
          "entity_ids",
          "community_ids",
          "claim_ids"
        ],
        "title": "Answer",
        "type": "object"
      },
...
      "Community": {
        "description": "A Leiden cluster of entities; derived data, replaced on every rebuild.\n\nThe report fields stay empty until a report has been generated.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "entity_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "minItems": 2,
            "title": "Entity Ids",
            "type": "array"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "title": "Findings",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "entity_ids"
        ],
        "title": "Community",
        "type": "object"
      },
      "Contributor": {
        "description": "A human or agent identity. Agents may propose and dispute only (ADR-0002).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "permissions": {
            "items": {
              "enum": [
                "propose",
                "review",
                "admin"
              ],
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "type": {
            "enum": [
              "human",
              "agent"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "display_name"
        ],
        "title": "Contributor",
        "type": "object"
      },
      "ContributorProfile": {
        "properties": {
          "accepted_claims": {
            "title": "Accepted Claims",
            "type": "integer"
          },
          "contributor": {
            "$ref": "#/components/schemas/Contributor"
          },
          "rejected_claims": {
            "title": "Rejected Claims",
            "type": "integer"
          }
        },
        "required": [
          "contributor",
          "accepted_claims",
          "rejected_claims"
        ],
        "title": "ContributorProfile",
        "type": "object"
      },
...
      "CreatedContributor": {
        "properties": {
          "api_key": {
            "description": "Shown only once; store it securely.",
            "title": "Api Key",
            "type": "string"
          },
          "contributor": {
            "$ref": "#/components/schemas/Contributor"
          }
        },
        "required": [
          "contributor",
          "api_key"
        ],
        "title": "CreatedContributor",
        "type": "object"
      },
...
      "Entity": {
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "embedding": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding"
          },
          "first_seen": {
            "format": "date-time",
            "title": "First Seen",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_updated": {
            "format": "date-time",
            "title": "Last Updated",
            "type": "string"
          },
          "merged_into_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merged Into Id"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "properties": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Properties",
            "type": "object"
          },
          "provenance": {
            "items": {
              "$ref": "#/components/schemas/ProvenanceRef"
            },
            "minItems": 1,
            "title": "Provenance",
            "type": "array"
          },
          "status": {
            "default": "active",
            "enum": [
              "active",
              "merged",
              "deprecated"
            ],
            "title": "Status",
            "type": "string"
          },
          "type": {
            "minLength": 1,
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type",
          "description",
          "confidence",
          "provenance"
        ],
        "title": "Entity",
        "type": "object"
      },
...
      "Finding": {
        "properties": {
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "text_unit_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "minItems": 1,
            "title": "Text Unit Ids",
            "type": "array"
          }
        },
        "required": [
          "text",
          "text_unit_ids"
        ],
        "title": "Finding",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JsonValue": {},
      "NeighborhoodOut": {
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/Entity"
          },
          "neighbors": {
            "items": {
              "$ref": "#/components/schemas/Entity"
            },
            "title": "Neighbors",
            "type": "array"
          },
          "relationships": {
            "items": {
              "$ref": "#/components/schemas/Relationship"
            },
            "title": "Relationships",
            "type": "array"
          },
          "text_units": {
            "items": {
              "$ref": "#/components/schemas/TextUnit"
            },
            "title": "Text Units",
            "type": "array"
          }
        },
        "required": [
          "entity",
          "neighbors",
          "relationships",
          "text_units"
        ],
        "title": "NeighborhoodOut",
        "type": "object"
      },
      "NewContributor": {
        "properties": {
          "display_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "permissions": {
            "items": {
              "enum": [
                "propose",
                "review",
                "admin"
              ],
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "type": {
            "enum": [
              "human",
              "agent"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "display_name"
        ],
        "title": "NewContributor",
        "type": "object"
      },
      "NewKey": {
        "properties": {
          "api_key": {
            "description": "Shown only once; the previous key stops working.",
            "title": "Api Key",
            "type": "string"
          }
        },
        "required": [
          "api_key"
        ],
        "title": "NewKey",
        "type": "object"
      },
...
      "Question": {
        "properties": {
          "mode": {
            "default": "local",
            "enum": [
              "local",
              "global",
              "hybrid"
            ],
            "title": "Mode",
            "type": "string"
          },
          "question": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "Question",
        "type": "object"
      },
      "RebuildStatus": {
        "properties": {
          "status": {
            "enum": [
              "not_started",
              "running",
              "failed",
              "complete"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "RebuildStatus",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Aether",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v0/communities": {
      "get": {
        "operationId": "list_communities_api_v0_communities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Community"
                  },
                  "title": "Response List Communities Api V0 Communities Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "description": "Database unavailable"
          }
        },
        "summary": "List Communities"
      }
    },
    "/api/v0/communities/rebuild": {
      "get": {
        "operationId": "get_rebuild_api_v0_communities_rebuild_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RebuildStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Rebuild"
      },
      "post": {
        "description": "Replace all communities and their reports. Queued behind any running extraction.",
        "operationId": "start_rebuild_api_v0_communities_rebuild_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RebuildStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Rebuild"
      }
    },
    "/api/v0/contributors": {
      "post": {
        "operationId": "create_contributor_api_v0_contributors_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewContributor"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedContributor"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Contributor"
      }
    },
    "/api/v0/contributors/me": {
      "get": {
        "operationId": "me_api_v0_contributors_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contributor"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Me"
      }
    },
    "/api/v0/contributors/me/key": {
      "post": {
        "description": "Replace your own API key, e.g. after a leak.",
        "operationId": "rotate_key_api_v0_contributors_me_key_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewKey"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "summary": "Rotate Key"
      }
    },
    "/api/v0/contributors/{contributor_id}": {
      "get": {
        "description": "Identity plus reputation: counts of this contributor's accepted and rejected claims.",
        "operationId": "profile_api_v0_contributors__contributor_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "contributor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Contributor Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContributorProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Profile"
      }
    },
    "/api/v0/contributors/{contributor_id}/revoke": {
      "post": {
        "description": "Disable a contributor's key; their past work stays attributed.",
        "operationId": "revoke_key_api_v0_contributors__contributor_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "contributor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Contributor Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Key"
      }
    },
...
    "/api/v0/entities": {
      "get": {
        "operationId": "search_entities_api_v0_entities_get",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Entity"
                  },
                  "title": "Response Search Entities Api V0 Entities Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Database unavailable"
          }
        },
        "summary": "Search Entities"
      }
    },
    "/api/v0/entities/{entity_id}/neighborhood": {
      "get": {
        "description": "The entity, its relationships and their other ends, and every cited text unit.",
        "operationId": "get_neighborhood_api_v0_entities__entity_id__neighborhood_get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeighborhoodOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Database unavailable"
          }
        },
        "summary": "Get Neighborhood"
      }
    },
    "/api/v0/query": {
      "post": {
        "description": "Answer with verified quotes. Modes: local (entities and their verified\nclaims), global (community reports), or hybrid (both).",
        "operationId": "query_api_v0_query_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Question"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too Many Requests"
          },
          "502": {
            "description": "Model provider unavailable"
          }
        },
        "summary": "Query"
      }
    },
...
  }
}